# cedar-terminology-server-application/src/main/resources/assets/swagger-api/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "CEDAR Terminology Server API",
    "title" : "CEDAR Terminology Server API",
    "version" : "2.9.10-SNAPSHOT"
  },
  "paths" : {
...
    "/bioportal/vs-collections/{vs_collection}/value-sets/{vs}/values" : {
      "get" : {
        "description" : "Find all values in a value set (either regular or provisional).",
        "operationId" : "findValuesByValueSet",
        "parameters" : [ {
          "description" : "Value set collection. Example: CEDARVS.",
          "in" : "path",
          "name" : "vs_collection",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Value set identifier. Example: http://www.semanticweb.org/jgraybeal/ontologies/2015/7/cedarvaluesets#Study_File_Type",
          "in" : "path",
          "name" : "vs",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Page to be returned. Example: 7.",
          "in" : "query",
          "name" : "page",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1
          }
        }, {
          "description" : "Number of results per page. Example: 10.",
          "in" : "query",
          "name" : "pageSize",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "description" : "Alias for the page size, accepted in either spelling.",
          "in" : "query",
          "name" : "page_size",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedValues"
                }
              }
            },
            "description" : "One page of the value set's values"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Not found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Find all values in a value set",
        "tags" : [ "Values" ]
      }
    },
    "/bioportal/vs-collections/{vs_collection}/values/{id}" : {
      "get" : {
        "description" : "Find value by id.",
        "operationId" : "findValue",
        "parameters" : [ {
          "description" : "Value identifier. Example: 42f22880-b04b-0133-848f-005056010073",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Value set collection. Example: CEDARVS.",
          "in" : "path",
          "name" : "vs_collection",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Value"
                }
              }
            },
            "description" : "The value"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Not found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Find value by id",
        "tags" : [ "Values" ]
      }
    },
    "/bioportal/vs-collections/{vs_collection}/values/{id}/all-values" : {
      "get" : {
        "description" : "Find all values in the value set that the given value belongs to.",
        "operationId" : "findAllValuesInValueSetByValue",
        "parameters" : [ {
          "description" : "Value identifier. Example: 42f22880-b04b-0133-848f-005056010073",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Value set collection. Example: CEDARVS.",
          "in" : "path",
          "name" : "vs_collection",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Page to be returned. Example: 7.",
          "in" : "query",
          "name" : "page",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1
          }
        }, {
          "description" : "Number of results per page. Example: 10.",
          "in" : "query",
          "name" : "pageSize",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "description" : "Alias for the page size, accepted in either spelling.",
          "in" : "query",
          "name" : "page_size",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedValues"
                }
              }
            },
            "description" : "One page of the values in the value set the given value belongs to"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Not found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Find all values in the value set that the given value belongs to",
        "tags" : [ "Values" ]
      }
    },
    "/bioportal/vs-collections/{vs_collection}/values/{id}/tree" : {
      "get" : {
        "description" : "Get value tree (only for regular values).",
        "operationId" : "findValueTree",
        "parameters" : [ {
          "description" : "Value identifier. Example: 42f22880-b04b-0133-848f-005056010073",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Value set collection. Example: CEDARVS.",
          "in" : "path",
          "name" : "vs_collection",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TreeNode"
                }
              }
            },
            "description" : "The value and the values around it in its value set"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Not found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Get value tree",
        "tags" : [ "Values" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CedarError" : {
        "type" : "object",
        "additionalProperties" : true,
        "description" : "A CEDAR error response. Diagnostic fields are populated only when they are relevant to the failure.",
        "properties" : {
          "entities" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Named domain entities associated with exception-mapped errors.",
              "nullable" : true
            },
            "description" : "Named domain entities associated with exception-mapped errors.",
            "nullable" : true
          },
          "errorId" : {
            "type" : "string",
            "format" : "uuid",
            "description" : "Correlation identifier for an internal error recorded in server logs.",
            "nullable" : true
          },
          "errorKey" : {
            "type" : "string",
            "description" : "Stable machine-readable error identifier.",
            "enum" : [ "templateElementNotCreated", "templateElementNotFound", "templateElementNotDeleted", "templateElementNotUpdated", "templateElementsNotListed", "templateFieldNotCreated", "templateFieldNotFound", "templateFieldNotDeleted", "templateFieldNotUpdated", "templateFieldsNotListed", "templateNotCreated", "templateNotFound", "templateNotDeleted", "templateNotUpdated", "templatesNotListed", "templateInstanceNotCreated", "templateInstanceNotFound", "templateInstanceNotDeleted", "templateInstanceNotUpdated", "templateInstancesNotListed", "noReadAccessToResource", "noWriteAccessToResource", "noReadAccessToFolder", "noWriteAccessToFolder", "noReadAccessToTemplate", "noWriteAccessToTemplate", "noReadAccessToTemplateElement", "noWriteAccessToTemplateElement", "noReadAccessToTemplateField", "noWriteAccessToTemplateField", "noReadAccessToTemplateInstance", "noWriteAccessToTemplateInstance", "noReadAccessToArtifact", "noWriteAccessToArtifact", "noReadAccessToCategory", "noWriteAccessToCategory", "folderNotFound", "artifactNotFound", "artifactPreconditionRequired", "artifactHasMovedOn", "verbatimWriteRefused", "nodeNotFound", "unknownResourceType", "missingParameter", "invalidInput", "sourceFolderNotFound", "sourceResourceNotFound", "targetFolderNotFound", "nodeNotMoved", "groupNotFound", "userNotFound", "uniqueConstraintCollision", "invalidData", "notAuthorized", "permissionMissing", "tokenInvalid", "userInfoLoadByTokenFailed", "userInfoLoadByApiKeyFailed", "cedarUserNotFound", "authorizationNotFound", "permissionNotOwned", "authorizationTypeUnknown", "apiKeyNotFound", "tokenMissing", "tokenExpired", "parentFolderNotSpecified", "parentFolderSpecifiedTwice", "pathNotNormalized", "parentFolderNotFound", "updateInvalidFolderName", "createInvalidFolderName", "nodeAlreadyPresent", "folderNotCreated", "missingNameAndDescription", "missingData", "folderNotDeleted", "artifactNotDeleted", "resourceNotCreated", "invalidResourceType", "invalidArtifactType", "readOtherProfileForbidden", "updateOtherProfileForbidden", "folderCanNotBeDeleted", "folderCanNotBeChanged", "groupAlreadyPresent", "groupCanBeModifiedOnlyByGroupAdmin", "groupCanBeDeletedOnlyByGroupAdmin", "groupMembersCanBeReadOnlyByGroupAdmin", "groupRequiresAdministrator", "groupUsersNotUpdated", "specialGroupCanNotBeDeleted", "folderPermissionsCanNotBeChanged", "unknownInstanceOutputFormat", "folderCopyNotAllowed", "methodNotImplemented", "upstreamServerError", "nothingToDo", "parentCategoryNotFound", "categoryAlreadyPresent", "categoryNotFound", "categoryCanNotBeDeleted", "rootCategoryCanNotBeDeleted", "unableToAttachCategory", "unableToDetachCategory", "noCategoriesWereAttached", "malformedJsonRequestBody", "malformedSearchTerm", "publishedArtifactCanNotBeChanged", "publishedArtifactCanNotBeDeleted", "versioningOnlyOnLatest", "versioningOnlyByOwner", "nonVersionedArtifactType", "createDraftOnlyFromPublished", "publishOnlyDraft", "draftNotCreated", "contentNotValid", "doiNotSupportedByResourceType", "doiCanNotBeSetForEmptyAtId", "doiCanNotBeAltered", "doiCanNotBeSet", "doiAlreadyExists", "dataCiteDOIDisabled", "resourceNotFound", "internalError" ],
            "nullable" : true
          },
          "errorMessage" : {
            "type" : "string",
            "description" : "Alias of message retained for compatibility.",
            "nullable" : true
          },
          "errorReasonKey" : {
            "type" : "string",
            "description" : "More specific machine-readable reason for the error.",
            "enum" : [ "nonEmptyFolder", "nonEmptyCategory", "userHomeFolder", "systemFolder", "templateReferencedInInstances", "validationError", "continuationExpired" ],
            "nullable" : true
          },
          "errorType" : {
            "type" : "string",
            "description" : "Broad category of the error.",
            "enum" : [ "notFound", "invalidArgument", "authorization", "permission", "server", "validationError" ],
            "nullable" : true
          },
          "message" : {
            "type" : "string",
            "description" : "Human-readable error message.",
            "nullable" : true
          },
          "objects" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Named structured values associated with the error."
            },
            "description" : "Named structured values associated with the error."
          },
          "operation" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Operation that failed. Its fields depend on the operation type.",
              "nullable" : true
            },
            "description" : "Operation that failed. Its fields depend on the operation type.",
            "nullable" : true
          },
          "parameters" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Named scalar values associated with the error."
            },
            "description" : "Named scalar values associated with the error."
          },
          "status" : {
            "type" : "string",
            "description" : "Symbolic HTTP response status.",
            "enum" : [ "BAD_REQUEST", "UNAUTHORIZED", "FORBIDDEN", "NOT_FOUND", "METHOD_NOT_ALLOWED", "NOT_ACCEPTABLE", "CONFLICT", "PRECONDITION_FAILED", "UNSUPPORTED_MEDIA_TYPE", "UNPROCESSABLE_ENTITY", "PRECONDITION_REQUIRED", "INTERNAL_SERVER_ERROR", "NOT_IMPLEMENTED", "BAD_GATEWAY", "SERVICE_UNAVAILABLE", "HTTP_VERSION_NOT_SUPPORTED" ]
          },
          "statusCode" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Numeric HTTP response status.",
            "maximum" : 599,
            "minimum" : 400
          },
          "suggestedAction" : {
            "type" : "string",
            "description" : "Recovery action suggested to the client.",
            "enum" : [ "none", "requestRole", "logout", "logoutImmediately", "provideAuthorizationHeader", "refreshToken" ],
            "nullable" : true
          }
        },
        "required" : [ "status", "statusCode" ]
      },
...
      "PagedValues" : {
        "type" : "object",
        "description" : "One page of value-set values.",
        "properties" : {
          "collection" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Value"
            }
          },
          "nextPage" : {
            "type" : "integer",
            "format" : "int32"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          },
          "prevPage" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalCount" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "Relation" : {
        "type" : "object",
        "properties" : {
          "@id" : {
            "type" : "string"
          },
          "@type" : {
            "type" : "string"
          },
          "created" : {
            "type" : "string"
          },
          "creator" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "relationType" : {
            "type" : "string"
          },
          "sourceClassId" : {
            "type" : "string"
          },
          "targetClassId" : {
            "type" : "string"
          },
          "targetClassOntology" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        }
      },
...
      "TreeNode" : {
        "type" : "object",
        "properties" : {
          "@id" : {
            "type" : "string"
          },
          "@type" : {
            "type" : "string"
          },
          "children" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TreeNode"
            }
          },
          "hasChildren" : {
            "type" : "boolean"
          },
          "id" : {
            "type" : "string"
          },
          "obsolete" : {
            "type" : "boolean"
          },
          "ontology" : {
            "type" : "string"
          },
          "prefLabel" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        }
      },
...
      "Value" : {
        "type" : "object",
        "properties" : {
          "@id" : {
            "type" : "string"
          },
          "@type" : {
            "type" : "string"
          },
          "created" : {
            "type" : "string"
          },
          "creator" : {
            "type" : "string"
          },
          "definitions" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "id" : {
            "type" : "string"
          },
          "notation" : {
            "type" : "string"
          },
          "prefLabel" : {
            "type" : "string"
          },
          "provisional" : {
            "type" : "boolean"
          },
          "relatedMatch" : {
            "type" : "string"
          },
          "relations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Relation"
            }
          },
          "synonyms" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "type" : {
            "type" : "string"
          },
          "vsCollection" : {
            "type" : "string"
          },
          "vsId" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "api_key" : {
        "description" : "Your CEDAR API key, using the format 'apiKey <yourApiKey>'",
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      }
    }
  }
}